% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdf4.R
\name{read_hdf4}
\alias{read_hdf4}
\title{Read HDF4 file}
\usage{
read_hdf4(file, ...)
}
\arguments{
\item{file}{(optional) path or source string to GDAL-readable raster}

\item{...}{ignored}
}
\value{
\code{rgdal SpatialGridDataFrame}
}
\description{
This function expects to use the HDF4Image driver of GDAL, and warns
if that's not the case.
}
\examples{
x <- read_hdf4()
print(class(x))
}
